Implement OperationCustomizer for custom annotation processing - Intercept all controller methods that have @DocumentedRequestBody or @Response annotation (or an abstraction of the latter) - Modify operation request and/or response(s) based on annotation contents - Remove ResponseCustomizer since both request and response need to be customized
